Start of Adding/Managing Weapons
Added in simple character info screen w/ level, class, weapon. Looks a
mess, will fix it later.

Fixed the DB creation so it would stop crashing (I did not know this,
but spaces and commas and parentheses in strings are totes important)

Added in new GetWeaponName function to the DB handler. Do NOT like this.
I shouldn't have to, but I can't get it to work any other way tonight.
Will look at it tomorrow.

The Attack Roll button on the main page now goes to my Basic Info page
because it wouldn't let me add a new button for some reason and I got
tired of fighting with the stupid thing.